server/requirements.txt
fastapi[all]==0.104.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-dotenv==1.0.1

flake8==6.1.0
isort==5.12.0